Separete admin/products tests into separate files
